+++ src/worker.ts
@@ -7,7 +7,7 @@ import {ApplicationUpdateAgent} from './agents/ApplicationUpdateAgent';
 import {BlockDispatcherWorker} from './workers/BlockDispatcherWorker';
 import {DeviationLeaderWorker} from "./workers/DeviationLeaderWorker";
 import {DeviationDispatcherWorker} from "./workers/DeviationDispatcherWorker";
-import {ValidatorListWorker} from "./workers/ValidatorListWorker";
+import {BlockchainMetricsWorker} from "./workers/BlockchainMetricsWorker";
 
 (async () => {
   await boot();
@@ -28,7 +28,7 @@ import {ValidatorListWorker} from "./workers/ValidatorListWorker";
     }
     case 'MetricsWorker': {
       Application.get(MetricsWorker).start();
-      Application.get(ValidatorListWorker).start();
+      Application.get(BlockchainMetricsWorker).start();
       break;
     }
     case 'DeviationLeaderWorker': {