+++ src/workers/BlockchainMetricsWorker.ts
@@ -3,12 +3,15 @@ import {inject, injectable} from 'inversify';
 
 import BasicWorker from './BasicWorker';
 import {ValidatorsResolver} from '../services/ValidatorsResolver';
+import {RequiredSignaturesResolver} from '../services/RequiredSignaturesResolver';
 
 @injectable()
-export class ValidatorListWorker extends BasicWorker {
+export class BlockchainMetricsWorker extends BasicWorker {
   @inject(ValidatorsResolver) validatorsResolver!: ValidatorsResolver;
+  @inject(RequiredSignaturesResolver) requiredSignaturesResolver!: RequiredSignaturesResolver;
+
   enqueue = async <T>(params: T, opts?: Bull.JobsOptions): Promise<Bull.Job<T> | undefined> => {
-    const isLocked = await this.connection.get(this.settings.jobs.validatorsResolver.lock.name);
+    const isLocked = await this.connection.get(this.settings.jobs.blockchainMetrics.lock.name);
     if (isLocked) return;
 
     return this.queue.add(this.constructor.name, params, opts);
@@ -17,18 +20,18 @@ export class ValidatorListWorker extends BasicWorker {
   apply = async (job: Bull.Job): Promise<void> => {
     if (this.isStale(job)) return;
 
-    const {lock} = this.settings.jobs.validatorsResolver;
+    const {lock} = this.settings.jobs.blockchainMetrics;
 
     const unlocked = await this.connection.set(lock.name, 'lock', 'EX', lock.ttl, 'NX');
 
     if (!unlocked) {
-      this.logger.error(`[ValidatorListWorker] apply for job but job !unlocked`);
+      this.logger.error(`[BlockchainMetricsWorker] apply for job but job !unlocked`);
       return;
     }
 
     try {
-      this.logger.debug(`[ValidatorListWorker] job run at ${new Date().toISOString()}`);
-      await this.validatorsResolver.apply();
+      this.logger.debug(`[BlockchainMetricsWorker] job run at ${new Date().toISOString()}`);
+      await Promise.all([this.validatorsResolver.apply(), this.requiredSignaturesResolver.apply()]);
     } catch (e) {
       this.logger.error(e);
     } finally {
@@ -38,7 +41,7 @@ export class ValidatorListWorker extends BasicWorker {
 
   isStale = (job: Bull.Job): boolean => {
     const age = new Date().getTime() - job.timestamp;
-    return age > this.settings.jobs.validatorsResolver.interval;
+    return age > this.settings.jobs.blockchainMetrics.interval;
   };
 
   start = (): void => {