+++ test/services/multichain/MultiChainStatusProcessor.test.ts
@@ -1,17 +1,24 @@
 import 'reflect-metadata';
 import {expect} from 'chai';
+import mongoose from "mongoose";
+import sinon from "sinon";
+import {BigNumber} from "ethers";
 
 import {getTestContainer} from '../../helpers/getTestContainer';
 import {MultiChainStatusProcessor} from '../../../src/services/multiChain/MultiChainStatusProcessor';
 import {ChainStatusWithAddress} from '../../../src/types/ChainStatus';
 import {chainStatusWithAddressFactory, chainStatusFactory} from '../../mocks/factories/chainStatusFactory';
 import {loadTestEnv} from "../../helpers/loadTestEnv";
-import mongoose from "mongoose";
+import {ValidatorRepository} from "../../../src/repositories/ValidatorRepository";
+import {Validator} from "../../../src/types/Validator";
+import {mockedLogger} from '../../mocks/logger';
+import SignatureCollector from "../../../src/services/SignatureCollector";
 
 describe('MultiChainStatusProcessor', () => {
+  const bscChainStatus = chainStatusWithAddressFactory.build();
+  let mockValidatorRepository: sinon.SinonStubbedInstance<ValidatorRepository>;
   let chainStatusWithAddress: ChainStatusWithAddress[];
   let multiChainStatusProcessor: MultiChainStatusProcessor;
-  const bscChainStatus = chainStatusWithAddressFactory.build();
 
   chainStatusWithAddress = [
     bscChainStatus,
@@ -29,12 +36,20 @@ describe('MultiChainStatusProcessor', () => {
 
   beforeEach(() => {
     const container = getTestContainer();
+
+    mockValidatorRepository = sinon.createStubInstance(ValidatorRepository);
+
+    container.rebind('Logger').toConstantValue(mockedLogger);
     container.bind(MultiChainStatusProcessor).toSelf();
+    container.bind(ValidatorRepository).toConstantValue(mockValidatorRepository);
+
     multiChainStatusProcessor = container.get(MultiChainStatusProcessor);
   });
 
   describe('when all chains can mint', () => {
     it('returns the chainsIdsReadyForBlock for all chains', async () => {
+      mockValidatorRepository.list.resolves([<Validator>{id: 'a', location: 'b', power: BigNumber.from(1)}]);
+
       const result = await multiChainStatusProcessor.apply(chainStatusWithAddress, 10);
 
       expect(result.chainsStatuses[0]).to.includes({chainId: 'bsc'});
@@ -61,6 +76,8 @@ describe('MultiChainStatusProcessor', () => {
     });
 
     it('returns the chainsIdsReadyForBlock for the one chain that can mint', async () => {
+      mockValidatorRepository.list.resolves([<Validator>{id: 'a', location: 'b', power: BigNumber.from(1)}]);
+
       const result = await multiChainStatusProcessor.apply(chainStatusWithAddress, 10);
       expect(result.chainsStatuses[0]).to.includes({chainId: 'bsc'});
       expect(result.chainsIdsReadyForBlock).to.deep.equal(['bsc']);