+++ CHANGELOG.md
@@ -6,8 +6,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## [7.11.0] - 2023-07-03
 ### Changed
 - pull validators list from target blockchain
+- do not overshoot heartbeat
+- cache required number of signatures
 
 ## [7.10.0] - 2023-06-15
 ### Added